# source/docs/swagger.yaml
basePath: /
definitions:
...
  web.ErrorResponse:
    properties:
      messages:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  web.RequestB2BPasien:
    properties:
      id:
        default: 0
        example: 1
        type: integer
      pasien_name:
        type: string
    required:
    - pasien_name
    type: object
...
  web.ResponseStandard:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/web.ErrorResponse'
        type: array
      fulfilled:
        type: integer
      message:
        type: string
      pagination:
        description: |-
          Meta       any `json:"meta"`
          JsonApi    any `json:"jsonapi"`
          Links      any `json:"links"`
          Included   any `json:"included"`
          Self       any `json:"self"`
          Related    any `json:"related"`
      statusCode:
        type: integer
    type: object
host: localhost:8200
info:
  contact: {}
  description: This is a sample server with v1 and v2 endpoints.
  title: Sarana Boilerplate API
  version: "2.0"
paths:
...
  /b2b-pasien:
    get:
      consumes:
      - application/json
      description: Get all b2b pasien
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Get all b2b pasien
      tags:
      - b2b-pasien
    post:
      consumes:
      - application/json
      description: Create b2b pasien
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.RequestB2BPasien'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Create b2b pasien
      tags:
      - b2b-pasien
  /b2b-pasien/{b2b_pasien_id}:
    delete:
      consumes:
      - application/json
      description: Delete b2b pasien
      parameters:
      - description: B2B pasien ID
        in: path
        name: b2b_pasien_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Delete b2b pasien
      tags:
      - b2b-pasien
    get:
      consumes:
      - application/json
      description: Get b2b pasien by ID
      parameters:
      - description: B2B pasien ID
        in: path
        name: b2b_pasien_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Get b2b pasien by ID
      tags:
      - b2b-pasien
    put:
      consumes:
      - application/json
      description: Update b2b pasien
      parameters:
      - description: B2B pasien ID
        in: path
        name: b2b_pasien_id
        required: true
        type: string
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.RequestB2BPasien'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Update b2b pasien
      tags:
      - b2b-pasien
  /b2b-pasien/name/{b2b_pasien_name}:
    get:
      consumes:
      - application/json
      description: Get b2b pasien by name
      parameters:
      - description: B2B pasien name
        in: path
        name: b2b_pasien_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Get b2b pasien by name
      tags:
      - b2b-pasien
...
swagger: "2.0"